// frontend/screens/air-quality-detailed-report/components/pollutant-info-card/pollutant-info-card.types.ts
import {Location} from '../../../../App.types.ts';
import {Pollutant} from '../../air-quality-detailed-report.types.ts';

export interface PollutantInfoCardProps {
    pollutantName: Pollutant;
    pollutantValue: number;
    pollutantDescription: string;
    selectedLocation?: Location;
    translatedName?: string;
    pollutantUnit?: string;
    viewHistoryText?: string;
}